% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moments.R
\name{AggregateRawMoments}
\alias{AggregateRawMoments}
\title{Aggregate Moments}
\usage{
AggregateRawMoments(momentsN, momentsY)
}
\arguments{
\item{momentsN}{a vector of raw moments of the claim frequency N.}

\item{momentsY}{a vector (possibly of the same length of momentsN) of raw moments of the claim intensities Y.}
}
\value{
The function returns the first raw moments of the aggregate distribution of S in the collective risk model.
}
\description{
the function returns the raw moments of the aggregate distribution of S in the collective risk model.
}
\note{
The raw moments of S are obtained using the probability generating function of S.
}
\examples{
AggregateRawMoments(momentsN=Moments("Poisson",10,"raw"),momentsY=Moments("Gamma",c(3,4),"raw"))
}
