% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moments.R
\name{centraltoraw}
\alias{centraltoraw}
\title{Function to convert Central Moments to raw ones.}
\usage{
centraltoraw(moments)
}
\arguments{
\item{moments}{is a vector of central moments. The first element must be the first raw moment and the nth element is the nth central moment.}
}
\value{
A vector of raw moments. the nth element of the vector
is the nth raw moment.
}
\description{
The function converts the central moments of a distribution to the raw ones.
}
\note{
Maximum length supported is n=5.
}
