% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approx.R
\name{FFT}
\alias{FFT}
\title{Fast Fourier Transform}
\usage{
FFT (phi, n,lower,upper, m )
}
\arguments{
\item{phi}{Characteristic function of S.}

\item{n}{number of points used in the discretization (possibly a power of 2)}

\item{lower}{starting point of the discretization of S}

\item{upper}{end point of the discretization of S}

\item{m}{approximation to P(S>m)}
}
\value{
the function returns P(S>m)
}
\description{
Computes the Fast Fourier Approximation to P(S>m), given the Characteristic function of S.
}
\details{
The function is useful when the characteristic function of a random variable is known, while the CDF is not.
this is the case in the collective risk model with light tailed claim intensities. For the case in which the intensities are heavy tailed,
refer to FFT2.
}
\examples{
FFT(phi= function(t) exp(30*((2/ (2- (0+1i)*t))^40 -1 )),n=2^19, lower=0,upper=3000,m=900)
}
\seealso{
FFT2
}
