% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approx.R
\name{FFT2}
\alias{FFT2}
\title{Fast Fourier Transform}
\usage{
FFT2 (fx,fm, n,lower,upper, m )
}
\arguments{
\item{fx}{CDF of the claim intesities.}

\item{n}{number of points used in the discretization (possibly a power of 2)}

\item{lower}{starting point of the discretization of X}

\item{upper}{end point of the discretization of X}

\item{m}{approximation to P(S>m)}

\item{fm}{Moment generating function of the Claim frequency.}
}
\value{
the function returns P(S>m)
}
\description{
Computes the Fast Fourier Approximation to P(S>m), where S is the aggregate claim size.
}
\details{
The function works for any aggregate claim distribution. It is not that stable when the lower and upper bounds are not selected with accuracy.
(be sure to consider a great part of the domain of X)
}
\note{
the discretization bounds (lower and upper) are generally different from those of FFT. In FFT the discretization
regards S, in FFT2 it concerns X.
}
\examples{
FFT2(fx=function(x) pgamma(x,40,2),fM= function(t) exp(30*(exp(t)-1)),n=2^19, lower=0, upper=3000,m=900)
}
\references{
"Panjer Recursion versus FFT For Compound Distributions", Paul Embrechets and Marco Frei (2009), Mathematical Methods of Operations Research.
}
\seealso{
FFT
}
