% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualRisk.R
\name{IndividualAggregateDist}
\alias{IndividualAggregateDist}
\title{Aggregate distribution in the Individual Risk Model}
\usage{
IndividualAggregateDist(x, n, dist, par)
}
\arguments{
\item{x}{the independent variable P(S<x).}

\item{n}{The number of policies}

\item{dist}{the distribution of each single policy (only "Gamma" and "InvGauss" are supported)}

\item{par}{the parameters of the distribution of each single policy.}
}
\value{
The function returns the cumulative distribution function of the aggregate claim size P(S<x)
}
\description{
The function returns the cumulative distribution function of the aggregate claim size.
}
\note{
Other distributions than the Gamma and the Inverse Gaussian might be considered. Nonetheless
It might be necessary to use convolutions or approximation methods to compute the CDF.
}
\examples{
1-IndividualAggregateDist(100,10,dist="InvGauss",par= c( 0.05,2.5))
}
