% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualRisk.R
\name{IndividualAggregateDist2}
\alias{IndividualAggregateDist2}
\title{Aggregate distribution in the Individual Risk Model}
\usage{
IndividualAggregateDist2(x, n, dist, par, q)
}
\arguments{
\item{x}{the independent variable P(S<x).}

\item{n}{The number of policies}

\item{dist}{the distribution of each single policy (only "Gamma" and "InvGauss" are supported)}

\item{par}{the parameters of the distribution of each single policy.}
}
\value{
The function returns the cumulative distribution function of the aggregate claim size P(S<x)
}
\description{
The function returns the cumulative distribution function of the aggregate claim size.
in this model specification there are n policies X and each policy is in the form X=I*B. Where I is the
indicator function (bernoulli(q)) and B is either a Gamma or an Inverse Gaussian random variable
(describing claim intensity given occurrence).
}
\note{
Other distributions than the Gamma and the Inverse Gaussian might be considered. Nonetheless
It might be necessary to use convolutions or approximation methods to compute the CDF.
}
\examples{
IndividualAggregateDist2(50,200,dist="Gamma",par=c(1,0.5),q=0.1)
}
