% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moments.R
\name{Moments}
\alias{Moments}
\title{Raw Moments of common distribution functions}
\usage{
Moments(dist, par, type)
}
\arguments{
\item{dist}{distribution for which you want to compute the moments: c("Binomial","Negbinomial"
,"Geometric", "Poisson", "Pareto", "Gamma", "Exponential", "LogNormal")}

\item{par}{vector of parameters of the chosen distribution.}

\item{type}{the function can return c("central", "raw", "standardized") moments.}
}
\value{
The  function returns the first 5 moments of the distribution chosen.
}
\description{
The function computes the first five moments of a distribution among the ones supported.
}
\note{
The function returns only the finite moments.
}
