% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approx.R
\name{MomentsApprox}
\alias{MomentsApprox}
\title{Aggregate distribution approximation based on moments}
\usage{
MomentsApprox(m,moments,type,n)
}
\arguments{
\item{m}{probability to approximate:  P(S>m)}

\item{moments}{vector of raw moments of the aggregate distribution S. Besides the "Normal" approximation, all other}

\item{type}{the available approximation methods are c("Normal","Gamma", "NormalPower", "NormalToGamma","GramCharlier", "Bowers")}

\item{n}{approximation order, only needed for GramCharlier and Bowers c(3,4,5).}

\item{approximations}{require at least the first 3 raw moments of S. "GramCharlier" and "Bowers" require as many raw moments
as their chosen approximation order.}
}
\description{
Function to approximate the Cumulative Distribution Function (CDF) of the aggregate claim size distribution.
}
\details{
The function takes as input the first raw moments of the aggregate claim size distribution S and returns
the desired approximation to P(S>m) based on moments. The available approximations are c("Normal","Gamma", "NormalPower", "NormalToGamma","GramCharlier", "Bowers").
For GramCharlier and Bowers, the approximations are available up to the 5th order (n=c(3,4,5))
}
\examples{
MomentsApprox(900,moments=AggregateRawMoments(Moments("Poisson",30,"raw"),Moments("Gamma",c(40,2),"raw")), "Bowers",n=3)
MomentsApprox(900,moments=AggregateRawMoments(Moments("Poisson",30,"raw"),Moments("Gamma",c(40,2),"raw")), "NormalToGamma")
}
