% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approx.R
\name{panjer}
\alias{panjer}
\title{Panjer approximation}
\usage{
panjer(fx, dist, par, step, m)
}
\arguments{
\item{fx}{continuous Probability Density Function (PDF) for the claim intensities.}

\item{dist}{c("Poisson","NegBin", "Binomial")}

\item{par}{vector of parameters of the claim frequency distributions (as ordered as in the built in PDF function)}

\item{step}{width of the discretization (for a computing time lower than a couple of minutes choose step such that m/step <20000)}

\item{m}{probability to approximate P(S>m)}
}
\value{
the function returns the upper and lower bound probabilities. The thinner the grid (lower the step) the closer the two bounds will get.
}
\description{
Computes the Panjer approximation to P(S>m) where S is the aggregate claim distribution resulting
from a claim frequency distribution in the (a,b,0) class and any continuous claim intensity distribution.
}
\examples{
panjer(fx=function(x) pgamma(x,40,2), dist="Poisson", par=30, step=0.1, 900)
}
