% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moments.R
\name{rawtocentral}
\alias{rawtocentral}
\title{Function to convert raw Moments to Central ones.}
\usage{
rawtocentral(moments)
}
\arguments{
\item{moments}{is a vector of raw moments}
}
\value{
A vector of central moments (the first one is the first raw moment.) the nth element of the vector
is the nth central moment.
}
\description{
The function converts the raw moments of a distribution to the central ones.
Instead of the first central moment (which is always 0), the function displays the first raw moment instead.
}
\note{
Maximum length supported is n=5.
}
