% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approx.R
\name{simulator}
\alias{simulator}
\title{Simulator of random numbers from some not so common distributions}
\usage{
simulator(dist, par, n)
}
\arguments{
\item{n}{the number of simulations.}

\item{distN}{For now distN= c("NegBin","Binomial","Poisson")}

\item{distY}{For now distY= c("Gamma", "Exponential","LogNormal","Weibul")}

\item{parN}{the parameters of the chosen counting distribution}

\item{parY}{the parameters of the chosen claim intensities distribution.}
}
\value{
The function returns the simulations of S. If you want to compute the probability P(S<m)
you can simply run the following command mean(S<m)
}
\description{
The function simulates random numbers from some distributions
}
\note{
Confidence intervals still to be added (along with more distributions, and truncated versions as well)
}
